+++ src/PyodideWorker.js
@@ -10,7 +10,6 @@ const PyodideWorker = () => {
   importScripts("https://cdn.jsdelivr.net/pyodide/v0.26.2/full/pyodide.js");
 
   const supportsAllFeatures = typeof SharedArrayBuffer !== "undefined";
-  console.log(`crossOriginIsolated: ${globalThis.crossOriginIsolated}`);
 
   // eslint-disable-next-line no-restricted-globals
   if (!supportsAllFeatures && name !== "incremental-features") {
@@ -367,6 +366,6 @@ globalThis.PyodideWorker = PyodideWorker;
 
 if (typeof module !== "undefined") {
   module.exports = {
-    PyodideWorker: PyodideWorker,
+    PyodideWorker,
   };
 }